1.修改用户获取 2.第一次取得token进行保存

+++ wenlibackyard/src/main/java/com/bean/AccessToken.java
@@ -10,17 +10,17 @@ public class AccessToken {
 	// 获取到的凭证
 	private String token;
 	// 凭证有效时间，单位：秒
-	private int expiresIn;
+	private long expiresIn;
 	public String getToken() {
 		return token;
 	}
 	public void setToken(String token) {
 		this.token = token;
 	}
-	public int getExpiresIn() {
+	public long getExpiresIn() {
 		return expiresIn;
 	}
-	public void setExpiresIn(int expiresIn) {
+	public void setExpiresIn(long expiresIn) {
 		this.expiresIn = expiresIn;
 	}
